solver: rename solver-next to solver
