+++ cache/remotecache/v1/cachestorage.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/moby/buildkit/identity"
-	solver "github.com/moby/buildkit/solver-next"
+	"github.com/moby/buildkit/solver"
 	"github.com/moby/buildkit/worker"
 	digest "github.com/opencontainers/go-digest"
 	"github.com/pkg/errors"
