+++ cache/remotecache/v1/chains.go
@@ -4,7 +4,7 @@ import (
 	"time"
 
 	"github.com/containerd/containerd/content"
-	solver "github.com/moby/buildkit/solver-next"
+	"github.com/moby/buildkit/solver"
 	digest "github.com/opencontainers/go-digest"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 )
