+++ cache/remotecache/v1/chains_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	solver "github.com/moby/buildkit/solver-next"
+	"github.com/moby/buildkit/solver"
 	digest "github.com/opencontainers/go-digest"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/stretchr/testify/require"
