+++ cache/remotecache/v1/parse.go
@@ -3,7 +3,7 @@ package cacheimport
 import (
 	"encoding/json"
 
-	solver "github.com/moby/buildkit/solver-next"
+	"github.com/moby/buildkit/solver"
 	"github.com/moby/buildkit/util/contentutil"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/pkg/errors"
