+++ cache/remotecache/v1/utils.go
@@ -5,7 +5,7 @@ import (
 	"sort"
 
 	"github.com/containerd/containerd/content"
-	solver "github.com/moby/buildkit/solver-next"
+	"github.com/moby/buildkit/solver"
 	digest "github.com/opencontainers/go-digest"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/pkg/errors"
