+++ solver/cache_test.go
