+++ solver/cachemanager.go
