+++ solver/cachestorage.go
