+++ solver/cachestorage_testsuite.go
