+++ solver/internal/pipe/pipe.go
