+++ solver/internal/pipe/pipe_test.go
