+++ solver/jobs.go
