+++ solver/memorycachestorage.go
