+++ solver/memorycachestorage_test.go
