+++ solver/progress.go
