+++ solver/result.go
