+++ solver/scheduler_test.go
