+++ solver/types.go
